Move k256 code from curve.rs to k256.rs
