+++ src/curve.rs
@@ -1,33 +1,17 @@
 //! Elliptic Curve abstraction
-use crate::{
-    errors::{
-        CallerError,
-        InternalError::{self, InternalInvariantFailed},
-        Result,
-    },
-    k256::{k256_order, K256},
-    p256::P256,
-};
+use crate::{errors::Result, k256::K256, p256::P256};
 use generic_array::GenericArray;
 use hmac::digest::core_api::CoreWrapper;
 use k256::{
-    ecdsa::{signature::DigestVerifier, VerifyingKey},
-    elliptic_curve::{
-        scalar::IsHigh, sec1::FromEncodedPoint, CurveArithmetic, Field, Group, PrimeField,
-    },
-    EncodedPoint, FieldBytes, ProjectivePoint, Scalar as K256_Scalar,
+    elliptic_curve::{scalar::IsHigh, Field, PrimeField},
+    FieldBytes, Scalar as K256_Scalar,
 };
 use libpaillier::unknown_order::BigNumber;
 use p256::Scalar as P256_Scalar;
 use serde::{Deserialize, Serialize};
-use sha2::Digest;
-use sha3::{Keccak256, Keccak256Core};
-use std::{
-    fmt::Debug,
-    ops::{Add, Deref},
-};
-use tracing::error;
-use zeroize::{Zeroize, Zeroizing};
+use sha3::Keccak256Core;
+use std::{fmt::Debug, ops::Add};
+use zeroize::Zeroize;
 
 /// Generic elliptic curve point.
 pub trait CT:
@@ -124,23 +108,6 @@ pub trait SignatureTrait: Clone + Copy + Debug + PartialEq {
     fn from_scalars(r: &BigNumber, s: &BigNumber) -> Result<Self>
     where
         Self: Sized + Debug;
-}
-
-/// ECDSA signature on a message.
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct CTSignatureK256<C: CT>(k256::ecdsa::Signature, std::marker::PhantomData<C>);
-impl<C: CT> CTSignatureK256<C> {
-    #[allow(dead_code)]
-    pub(crate) fn recovery_id(&self, message: &[u8], public_key: &VerifyingKey) -> Result<u8> {
-        let digest = Keccak256::new_with_prefix(message);
-        let recover_id =
-            k256::ecdsa::RecoveryId::trial_recovery_from_digest(public_key, digest, &self.0)
-                .map_err(|e| {
-                    error!("Failed to compute recovery ID for signature. Reason: {e:?}");
-                    CallerError::SignatureTrialRecoveryFailed
-                })?;
-        Ok(recover_id.into())
-    }
 }
 
 /// Verifying Key trait
@@ -319,156 +286,6 @@ pub type Secp256r1 = P256;
 
 /// P256 scalar type.
 pub type P256Scalar = P256_Scalar;
-
-impl SignatureTrait for CTSignatureK256<K256> {
-    fn from_scalars(r: &BigNumber, s: &BigNumber) -> Result<Self> {
-        let r_scalar = <K256 as CT>::bn_to_scalar(r)?;
-        let s_scalar = <K256 as CT>::bn_to_scalar(s)?;
-        let sig = k256::ecdsa::Signature::from_scalars(r_scalar, s_scalar)
-            .map_err(|_| InternalInvariantFailed)?;
-        Ok(CTSignatureK256(sig, std::marker::PhantomData::<K256>))
-    }
-}
-
-impl Deref for CTSignatureK256<K256> {
-    type Target = k256::ecdsa::Signature;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl CT for K256 {
-    const GENERATOR: Self = K256::GENERATOR;
-    const IDENTITY: Self = K256::IDENTITY;
-    type Scalar = K256_Scalar;
-    type Encoded = EncodedPoint;
-    type Projective = ProjectivePoint;
-    type VK = VerifyingKey;
-    type ECDSASignature = CTSignatureK256<K256>;
-
-    fn order() -> BigNumber {
-        k256_order()
-    }
-
-    fn mul_by_bn(&self, scalar: &BigNumber) -> Result<Self> {
-        self.multiply_by_bignum(scalar)
-    }
-
-    fn scale_generator(scalar: &BigNumber) -> Result<Self> {
-        K256::GENERATOR.multiply_by_bignum(scalar)
-    }
-
-    fn mul(&self, scalar: &Self::Scalar) -> Self {
-        self.multiply_by_scalar(scalar)
-    }
-
-    fn x_projection(&self) -> Result<Self::Scalar> {
-        let x_projection = self.x_affine();
-
-        // Note: I don't think this is a foolproof transformation. The `from_repr`
-        // method expects a scalar in the range `[0, q)`, but there's no
-        // guarantee that the x-coordinate of `R` will be in that range.
-        Option::from(<k256::Scalar as PrimeField>::from_repr(x_projection)).ok_or_else(|| {
-            error!("Unable to compute x-projection of curve point: failed to convert x coord to `Scalar`");
-            InternalInvariantFailed
-        })
-    }
-
-    fn to_bytes(self) -> Vec<u8> {
-        K256::to_bytes(self)
-    }
-
-    fn try_from_bytes_ct(bytes: &[u8]) -> Result<Self> {
-        K256::try_from_bytes(bytes)
-    }
-
-    // Returns x: BigNumber as a k256::Scalar mod k256_order
-    fn bn_to_scalar(x: &BigNumber) -> Result<Self::Scalar> {
-        // Take (mod q)
-        let order = Self::order();
-
-        let x_modded = x % order;
-
-        let bytes = Zeroizing::new(x_modded.to_bytes());
-        let mut slice = Zeroizing::new(vec![0u8; 32 - bytes.len()]);
-        slice.extend_from_slice(&bytes);
-
-        let mut ret: Self::Scalar = Option::from(<k256::Scalar as PrimeField>::from_repr(
-            GenericArray::clone_from_slice(&slice),
-        ))
-        .ok_or_else(|| {
-            error!("Failed to convert BigNumber into k256::Scalar");
-            InternalError::InternalInvariantFailed
-        })?;
-
-        // Make sure to negate the scalar if the original input was negative
-        if x < &BigNumber::zero() {
-            ret = ret.negate();
-        }
-
-        Ok(ret)
-    }
-
-    fn multiply_by_bignum(&self, point: &BigNumber) -> Result<Self> {
-        let s = Zeroizing::new(Self::bn_to_scalar(point)?);
-        let p = self.multiply_by_scalar(&s);
-        Ok(p)
-    }
-
-    fn multiply_by_scalar(&self, point: &Self::Scalar) -> Self {
-        self.multiply_by_scalar(point)
-    }
-
-    // Convert from k256::Scalar to BigNumber
-    fn scalar_to_bn(x: &Self::Scalar) -> BigNumber {
-        let bytes = x.to_repr();
-        BigNumber::from_slice(bytes)
-    }
-
-    // Random point.
-    fn random() -> Self {
-        let mut rng = rand::thread_rng();
-        let random_point = ProjectivePoint::random(&mut rng);
-        K256(random_point)
-    }
-}
-
-impl VKT for VerifyingKey {
-    type C = K256;
-
-    fn from_point(point: Self::C) -> Result<Self> {
-        VerifyingKey::from_sec1_bytes(&point.to_bytes()).map_err(|_| InternalInvariantFailed)
-    }
-
-    fn verify_signature(
-        &self,
-        digest: Keccak256,
-        signature: <Self::C as CT>::ECDSASignature,
-    ) -> Result<()> {
-        self.verify_digest(digest, signature.deref())
-            .map_err(|_| InternalInvariantFailed)
-    }
-
-    /// Add two verifying keys.
-    fn add(&self, other: &Self) -> Self {
-        //let point = self.to_encoded_point(false);
-        //let other_point = other.to_encoded_point(false);
-        //let sum = point.add(&other_point).unwrap();
-        //VerifyingKey::from_encoded_point(&sum).unwrap()
-        let point1 = self.to_encoded_point(false);
-        let point2 = other.to_encoded_point(false);
-        let p1 = ProjectivePoint::from_encoded_point(&point1)
-            .expect("Can not convert the first argument");
-        let p2 = ProjectivePoint::from_encoded_point(&point2)
-            .expect("Can not convert the second argument");
-        let sum = p1 + p2;
-        let sum_affine: ProjectivePoint =
-            <k256::Secp256k1 as CurveArithmetic>::AffinePoint::from(sum).into();
-        VerifyingKey::from_affine((&sum_affine).into())
-            .expect("Can not convert the sum to verifying key")
-    }
-}
 
 #[cfg(test)]
 mod tests {
